refactor: atualiza a estrutura após atualizar com as alterações da main
