+++ grails-app/conf/spring/resources.groovy
@@ -1,6 +1,7 @@
 package spring
 
-import core.listeners.UserPasswordEncoderListener
+import com.mini.asaas.user.UserPasswordEncoderListener
+
 // Place your Spring DSL code here
 beans = {
     userPasswordEncoderListener(UserPasswordEncoderListener)
