--- grails-app/domain/com/mini/asaas/Role.groovy
@@ -1,23 +1,0 @@
-package com.mini.asaas
-
-import groovy.transform.EqualsAndHashCode
-import groovy.transform.ToString
-import grails.compiler.GrailsCompileStatic
-
-@GrailsCompileStatic
-@EqualsAndHashCode(includes='authority')
-@ToString(includes='authority', includeNames=true, includePackage=false)
-class Role implements Serializable {
-
-	private static final long serialVersionUID = 1
-
-	String authority
-
-	static constraints = {
-		authority nullable: false, blank: false, unique: true
-	}
-
-	static mapping = {
-		cache true
-	}
-}
