+++ grails-app/domain/com/mini/asaas/user/Role.groovy
@@ -1,0 +1,24 @@
+package com.mini.asaas.user
+
+import grails.compiler.GrailsCompileStatic
+import groovy.transform.EqualsAndHashCode
+import groovy.transform.ToString
+
+@GrailsCompileStatic
+@EqualsAndHashCode(includes = "authority")
+@ToString(includes = "authority", includeNames = true, includePackage = false)
+class Role implements Serializable {
+
+    private static final long serialVersionUID = 1
+
+    String authority
+
+    static constraints = {
+        authority blank: false
+    }
+
+    static mapping = {
+        cache true
+        authority unique: true
+    }
+}
