+++ grails-app/domain/com/mini/asaas/user/User.groovy
@@ -1,13 +1,13 @@
-package com.mini.asaas
+package com.mini.asaas.user
 
-import core.entities.BaseEntity
+import com.mini.asaas.base.BaseEntity
+import grails.compiler.GrailsCompileStatic
 import groovy.transform.EqualsAndHashCode
 import groovy.transform.ToString
-import grails.compiler.GrailsCompileStatic
 
 @GrailsCompileStatic
-@EqualsAndHashCode(includes='email')
-@ToString(includes='email', includeNames=true, includePackage=false)
+@EqualsAndHashCode(includes = "email")
+@ToString(includes = "email", includeNames = true, includePackage = false)
 class User extends BaseEntity implements Serializable {
 
     private static final long serialVersionUID = 1
@@ -30,12 +30,12 @@ class User extends BaseEntity implements Serializable {
 
     static constraints = {
         password blank: false, password: true
-        email blank: false, unique: true, email: true
+        email blank: false, email: true
     }
 
     static mapping = {
         table '`user`'
-	    password column: '`password`'
-        email column: '`email`'
+        password column: '`password`'
+        email unique: true
     }
 }
